+++ .vscode/launch.json
@@ -389,6 +389,38 @@
                 }
             ]
         },{
+            "name": "Test NodeControllerSnapshot",
+            "type": "cppdbg",
+            "request": "launch",
+            "program": "${workspaceFolder}/build/k2eg-test",
+            "args": ["--gtest_filter=NodeControllerSnapshot.*"],
+            "stopAtEntry": false,
+            "cwd": "${workspaceFolder}",
+            "environment": [
+                // {
+                //     "name": "LSAN_OPTIONS",
+                //     "value": "verbosity=1:log_threads=1"
+                // },
+                {
+                    "name": "LD_LIBRARY_PATH",
+                    "value": "$LD_LIBRARY_PATH:${workspaceFolder}/build/local/lib:${workspaceFolder}/build/local/lib/linux_86_64"
+                }
+            ],
+            "externalConsole": false,
+            "MIMode": "gdb",
+            "setupCommands": [
+                {
+                    "description": "Enable pretty-printing for gdb",
+                    "text": "-enable-pretty-printing",
+                    "ignoreFailures": true
+                },
+                {
+                    "description": "Set Disassembly Flavor to Intel",
+                    "text": "-gdb-set disassembly-flavor intel",
+                    "ignoreFailures": true
+                }
+            ]
+        },{
             "name": "Test NodeControllerMonitorChecker",
             "type": "cppdbg",
             "request": "launch",
