+++ test/controller/NodeControllerCommon.h
@@ -1,24 +1,119 @@
 #ifndef NODECONTROLLERCOMMON_H_
 #define NODECONTROLLERCOMMON_H_
 
+#include <gtest/gtest.h>
+
 #include <latch>
 #include <string>
 
+#include <k2eg/common/utility.h>
 #include "k2eg/common/ProgramOptions.h"
 #include "k2eg/service/ServiceResolver.h"
 #include "k2eg/service/log/ILogger.h"
 #include "k2eg/service/log/impl/BoostLogger.h"
 #include "k2eg/service/pubsub/IPublisher.h"
+#include "boost/json/object.hpp"
+#include <k2eg/controller/node/NodeController.h>
+#include <k2eg/service/ServiceResolver.h>
+#include <k2eg/service/configuration/configuration.h>
+#include <k2eg/service/data/DataStorage.h>
+#include <k2eg/service/epics/EpicsServiceManager.h>
+#include <k2eg/service/metric/impl/prometheus/PrometheusMetricService.h>
+#include <k2eg/service/pubsub/pubsub.h>
+#include <k2eg/service/scheduler/Scheduler.h>
+#include <k2eg/service/metric/IMetricService.h>
+
+#include <filesystem>
 
 using namespace k2eg::common;
+
+using namespace k2eg::controller::node::configuration;
+using namespace k2eg::controller::node::worker::monitor;
+using namespace k2eg::controller::command;
+using namespace k2eg::controller::command::cmd;
+using namespace k2eg::controller::node;
+
 using namespace k2eg::service;
 using namespace k2eg::service::log;
 using namespace k2eg::service::log::impl;
 using namespace k2eg::service::pubsub;
-using namespace k2eg::controller::node::configuration;
-using namespace k2eg::controller::node::worker::monitor;
+using namespace k2eg::service;
 using namespace k2eg::service::data;
 using namespace k2eg::service::data::repository;
+using namespace k2eg::service::log;
+using namespace k2eg::service::log::impl;
+using namespace k2eg::service::pubsub;
+using namespace k2eg::service::data;
+using namespace k2eg::service::scheduler;
+using namespace k2eg::service::epics_impl;
+using namespace k2eg::service::configuration;
+using namespace k2eg::service::configuration::impl::consul;
+using namespace k2eg::service::metric;
+using namespace k2eg::service::metric::impl::prometheus_impl;
+namespace bs = boost::system;
+namespace bj = boost::json;
+namespace fs = std::filesystem;
+
+class DummyPublisherCounter : public IPublisher
+{
+    std::uint64_t counter;
+
+public:
+    std::latch l;
+    DummyPublisherCounter(unsigned int latch_counter)
+        : IPublisher(std::make_unique<const PublisherConfiguration>(PublisherConfiguration{.server_address = "fake_"
+                                                                                                             "addres"
+                                                                                                             "s"}))
+        , l(latch_counter)
+        , counter(0) {};
+    ~DummyPublisherCounter() = default;
+
+    void setAutoPoll(bool autopoll) {}
+
+    int setCallBackForReqType(const std::string req_type, EventCallback eventCallback)
+    {
+        return 0;
+    }
+
+    int createQueue(const QueueDescription& queue)
+    {
+        return 0;
+    }
+
+    int deleteQueue(const std::string& queue_name)
+    {
+        return 0;
+    }
+
+    QueueMetadataUPtr getQueueMetadata(const std::string& queue_name)
+    {
+        return nullptr;
+    }
+
+    int flush(const int timeo)
+    {
+        return 0;
+    }
+
+    int pushMessage(PublishMessageUniquePtr message, const PublisherHeaders& header = PublisherHeaders())
+    {
+        counter++;
+        l.count_down();
+        return 0;
+    }
+
+    int pushMessages(PublisherMessageVector& messages, const PublisherHeaders& header = PublisherHeaders())
+    {
+        counter += messages.size();
+        l.count_down(messages.size());
+        return 0;
+    }
+
+    size_t getQueueMessageSize()
+    {
+        return counter;
+    }
+};
 
 class ControllerConsumerDummyPublisher : public IPublisher
 {
@@ -200,4 +295,209 @@ public:
     }
 };
 
+class DummyPublisherNoSignal : public IPublisher
+{
+public:
+    std::vector<PublishMessageSharedPtr> sent_messages;
+    DummyPublisherNoSignal()
+        : IPublisher(std::make_unique<const PublisherConfiguration>(PublisherConfiguration{.server_address = "fake_"
+                                                                                                             "addres"
+                                                                                                             "s"})) {};
+    ~DummyPublisherNoSignal() = default;
+
+    void setAutoPoll(bool autopoll) {}
+
+    int setCallBackForReqType(const std::string req_type, EventCallback eventCallback)
+    {
+        return 0;
+    }
+
+    int createQueue(const QueueDescription& queue)
+    {
+        return 0;
+    }
+
+    int deleteQueue(const std::string& queue_name)
+    {
+        return 0;
+    }
+
+    QueueMetadataUPtr getQueueMetadata(const std::string& queue_name)
+    {
+        return nullptr;
+    }
+
+    int flush(const int timeo)
+    {
+        return 0;
+    }
+
+    int pushMessage(PublishMessageUniquePtr message, const PublisherHeaders& header = PublisherHeaders())
+    {
+        sent_messages.push_back(std::move(message));
+        return 0;
+    }
+
+    int pushMessages(PublisherMessageVector& messages, const PublisherHeaders& header = PublisherHeaders())
+    {
+        messages.clear();
+        return 0;
+    }
+
+    size_t getQueueMessageSize()
+    {
+        return sent_messages.size();
+    }
+};
+
+
+inline boost::json::object getJsonObject(PublishMessage& published_message)
+{
+    bs::error_code  ec;
+    bj::string_view value_str = bj::string_view(published_message.getBufferPtr(), published_message.getBufferSize());
+    auto            result = bj::parse(value_str, ec).as_object();
+    if (ec)
+        throw std::runtime_error("invalid json");
+    return result;
+}
+
+inline msgpack::unpacked getMsgPackObject(PublishMessage& published_message)
+{
+    msgpack::unpacked msg_upacked;
+    msgpack::unpack(msg_upacked, published_message.getBufferPtr(), published_message.getBufferSize());
+    return msg_upacked;
+}
+
+inline void wait_forPublished_message_size(DummyPublisherNoSignal& publisher, unsigned int requested_size, unsigned int timeout_ms)
+{
+    auto start_time = std::chrono::steady_clock::now();
+    auto end_time = std::chrono::steady_clock::now();
+    std::chrono::duration<long, std::milli> tout = std::chrono::duration_cast<std::chrono::milliseconds>(end_time - start_time);
+    bool waiting = true;
+    while (waiting)
+    {
+        waiting = publisher.getQueueMessageSize() < requested_size;
+        waiting = waiting && (tout.count() < timeout_ms);
+        std::this_thread::sleep_for(std::chrono::milliseconds(100));
+        end_time = std::chrono::steady_clock::now();
+        tout = std::chrono::duration_cast<std::chrono::milliseconds>(end_time - start_time);
+    }
+}
+
+
+inline boost::json::object exstractJsonObjectThatContainsKey(std::vector<PublishMessageSharedPtr>& messages, const std::string& key_to_find, const std::string& published_on_topic, bool log = false)
+{
+    for (int idx = 0; idx < messages.size(); idx++)
+    {
+        std::cout << "queue:" << messages[idx]->getQueue() << std::endl;
+        if (messages[idx]->getQueue().compare(published_on_topic) != 0)
+            continue;
+        auto json_obj = getJsonObject(*messages[idx]);
+        if (log)
+        {
+            std::cout << json_obj << std::endl;
+        }
+        if (json_obj.contains(key_to_find))
+        {
+            return json_obj;
+        }
+    }
+    return boost::json::object();
+}
+
+inline msgpack::unpacked exstractMsgpackObjectThatContainsKey(std::vector<PublishMessageSharedPtr>& messages, const std::string& key_to_find, const std::string& published_on_topic, bool log = false)
+{
+    typedef std::map<std::string, msgpack::object> Map;
+    typedef std::vector<msgpack::object>           Vec;
+    for (int idx = 0; idx < messages.size(); idx++)
+    {
+        if (messages[idx]->getQueue().compare(published_on_topic) != 0)
+            continue;
+        auto msgpack_obj = getMsgPackObject(*messages[idx]);
+        switch (msgpack_obj->type)
+        {
+        case msgpack::type::MAP:
+            {
+                auto map_reply = msgpack_obj->as<Map>();
+                if (map_reply.contains(key_to_find))
+                {
+                    return msgpack_obj;
+                }
+                break;
+            }
+
+        case msgpack::type::ARRAY:
+            {
+                return msgpack_obj;
+            }
+        }
+    }
+    return msgpack::unpacked();
+}
+
+inline std::size_t countMessageOnTopic(std::vector<PublishMessageSharedPtr>& messages, const std::string& published_on_topic)
+{
+    std::size_t counter = 0;
+    for (int idx = 0; idx < messages.size(); idx++)
+    {
+        if (messages[idx]->getQueue().compare(published_on_topic) != 0)
+            continue;
+        counter++;
+    }
+    return counter;
+}
+
+
+inline std::unique_ptr<NodeController> initBackend(int& tcp_port, IPublisherShrdPtr pub, bool enable_debug_log = false, bool reset_conf = true)
+{
+    int         argc = 1;
+    const char* argv[1] = {"epics-k2eg-test"};
+    clearenv();
+    if (enable_debug_log)
+    {
+        setenv("EPICS_k2eg_log-on-console", "true", 1);
+        setenv("EPICS_k2eg_log-level", "trace", 1);
+    }
+    else
+    {
+        setenv("EPICS_k2eg_log-on-console", "false", 1);
+    }
+
+    if (reset_conf)
+    {
+        setenv("EPICS_k2eg_configuration-reset-on-start", "true", 1);
+    }
+
+    setenv("EPICS_k2eg_metric-server-http-port", std::to_string(++tcp_port).c_str(), 1);
+    setenv(("EPICS_k2eg_" + std::string(SCHEDULER_CHECK_EVERY_AMOUNT_OF_SECONDS)).c_str(), "1", 1);
+    // set monitor expiration time out at minimum
+    setenv(("EPICS_k2eg_" + std::string(NC_MONITOR_EXPIRATION_TIMEOUT)).c_str(), "1", 1);
+    setenv(("EPICS_k2eg_" + std::string(CONFIGURATION_SERVICE_HOST)).c_str(), "consul", 1);
+    std::unique_ptr<ProgramOptions> opt = std::make_unique<ProgramOptions>();
+    opt->parse(argc, argv);
+    ServiceResolver<INodeConfiguration>::registerService(std::make_shared<ConsuleNodeConfiguration>(opt->getConfigurationServiceConfiguration()));
+    ServiceResolver<Scheduler>::registerService(std::make_shared<Scheduler>(opt->getSchedulerConfiguration()));
+    ServiceResolver<Scheduler>::resolve()->start();
+    ServiceResolver<ILogger>::registerService(std::make_shared<BoostLogger>(opt->getloggerConfiguration()));
+    ServiceResolver<IMetricService>::registerService(std::make_shared<PrometheusMetricService>(opt->getMetricConfiguration()));
+    ServiceResolver<EpicsServiceManager>::registerService(std::make_shared<EpicsServiceManager>());
+    ServiceResolver<IPublisher>::registerService(pub);
+    DataStorageUPtr storage = std::make_unique<DataStorage>(fs::path(fs::current_path()) / "test.sqlite");
+    // data should be alwasys erased becasue now the local databas eis only used at runtime the persistent
+    // data is stored on the central configuration management server
+    toShared(storage->getChannelRepository())->removeAll();
+    return std::make_unique<NodeController>(opt->getNodeControllerConfiguration(), std::move(storage));
+}
+
+inline void deinitBackend(std::unique_ptr<NodeController> node_controller)
+{
+    node_controller.reset();
+    EXPECT_NO_THROW(ServiceResolver<IPublisher>::resolve().reset(););
+    EXPECT_NO_THROW(ServiceResolver<EpicsServiceManager>::resolve().reset(););
+    EXPECT_NO_THROW(ServiceResolver<IMetricService>::resolve().reset(););
+    EXPECT_NO_THROW(ServiceResolver<ILogger>::resolve().reset(););
+    EXPECT_NO_THROW(ServiceResolver<Scheduler>::resolve()->stop(););
+    EXPECT_NO_THROW(ServiceResolver<Scheduler>::resolve().reset(););
+    EXPECT_NO_THROW(ServiceResolver<INodeConfiguration>::resolve().reset(););
+}
 #endif
